requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-tailwind==0.2.0
django-crispy-forms==1.10.0
idna==3.4
requests==2.30.0
urllib3==2.0.2